backend/requirements.txt
asgiref==3.6.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
defusedxml==0.8.0rc2
Django==4.1.7
django-dotenv==1.4.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
gindix-multitenant==3.1.1
idna==3.7
oauthlib==3.2.2
psycopg2==2.9.5
pycparser==2.22
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2022.7.1
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.4.3
tzdata==2024.1
urllib3==2.2.2
